Fix a warning with Visual C++
git-svn-id: svn://solarus-engine.org/solarus@327 08e91b48-8f19-0410-8fc9-9db6b7bce602

+++ src/SelectionMenu.cpp
@@ -998,7 +998,7 @@ void SelectionMenu::redraw_choose_name_screen(void) {
  */
 bool SelectionMenu::select_letter(void) {
 
-  int length = strlen(player_name);
+  size_t length = strlen(player_name);
   char letter_to_add = '\0';
   bool finished = false;
